05-data-platforms/my-taxi-pipeline/pipeline/assets/ingestion/requirements.txt
pandas==2.2.2
requests==2.32.3
pyarrow==18.1.0
python-dateutil==2.9.0.post0
